requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.1
djangorestframework==3.11.1
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
dj-rest-auth==1.1.1
drf-yasg==1.17.1
uritemplate==3.0.1
PyYAML==5.3.1
django-allauth==0.42.0
